+++ src/components/Confirm/Confirm.js
@@ -4,7 +4,7 @@ import { isOfMimeType } from '~utils/blob'
 import {
   uploadDocument,
   uploadLivePhoto,
-  uploadLiveVideo,
+  uploadFaceVideo,
   sendMultiframeSelfie,
 } from '~utils/onfidoApi'
 import { poaDocumentTypes } from '../DocumentSelector/documentTypes'
@@ -188,6 +188,17 @@ class Confirm extends Component {
     } = capture
     this.setState({ capture })
 
+    if (method === 'face') {
+      if (variant === 'video') {
+        const data = { challengeData, blob, language, sdkMetadata }
+        uploadFaceVideo(data, url, token, this.onApiSuccess, this.onApiError)
+        return
+      }
+
+      this.handleSelfieUpload(capture, token)
+      return
+    }
+
     if (method === 'document') {
       const isPoA = poaDocumentTypes.includes(poaDocumentType)
       const shouldPerformImageQualityValidations =
@@ -220,13 +231,6 @@ class Confirm extends Component {
         ...issuingCountry,
       }
       uploadDocument(data, url, token, this.onApiSuccess, this.onApiError)
-    } else if (method === 'face') {
-      if (variant === 'video') {
-        const data = { challengeData, blob, language, sdkMetadata }
-        uploadLiveVideo(data, url, token, this.onApiSuccess, this.onApiError)
-      } else {
-        this.handleSelfieUpload(capture, token)
-      }
     }
   }