+++ src/components/crossDevice/MobileConnected/index.js
@@ -1,7 +1,7 @@
 import { h } from 'preact'
 import PageTitle from '../../PageTitle'
 import { trackComponent } from '../../../Tracker'
-import { preventDefaultOnClick } from '~utils/index'
+import { preventDefaultOnClick } from '~utils'
 import { localised } from '../../../locales'
 import theme from '../../Theme/style.scss'
 import style from './style.scss'